# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SQL over HTTP
  version: 0.1.0
  description: Executes raw SQL against a SQLite database replicated to GCS with Litestream.
paths:
  /health:
    get:
      operationId: health
      summary: Check Litestream and database responsiveness
      responses:
        "204":
          description: Litestream and the database are responsive
        "503":
          description: Litestream or the database is unavailable
  /sql:
    post:
      operationId: executeSql
      description: |
        Executes the statements in order and returns one result per statement. Two or more
        statements run in a single transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/Statement"
            example:
              - sql: INSERT INTO users (id, name) VALUES (?, ?)
                params: [1, alice]
              - sql: SELECT id, name FROM users WHERE id = ?
                params: [1]
      responses:
        "200":
          description: Statements executed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
              example:
                - columns: []
                  types: []
                  rows: []
                  changes: 1
                - columns: [id, name]
                  types: [INTEGER, TEXT]
                  rows: [[1, alice]]
                  changes: 0
        "400":
          description: SQL error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: No database connection available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sync:
    post:
      operationId: syncReplica
      summary: Replicate to GCS
      responses:
        "204":
          description: All changes are replicated
        "500":
          description: Litestream sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The database was handed off to a new revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stop:
    post:
      operationId: stop
      description: Hands the database off to a new revision.
      parameters:
        - name: revision
          in: query
          required: true
          description: The caller's revision. An instance doesn't stop on behalf of its own revision.
          schema:
            type: string
      responses:
        "204":
          description: The database was handed off
        "409":
          description: The caller is this revision, so nothing was stopped
        "500":
          description: Litestream sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    Statement:
      type: object
      required: [sql]
      properties:
        sql:
          type: string
          description: A single SQL statement with `?` placeholders.
        params:
          type: array
          description: Values bound to the placeholders in order. Booleans are bound as 0 or 1.
          items:
            description: A JSON null, boolean, number or string.
    Result:
      type: object
      required: [columns, types, rows, changes]
      properties:
        columns:
          type: array
          items:
            type: string
        types:
          type: array
          description: Declared type of each column (e.g. `DATETIME`, `BOOLEAN`)
          items:
            type: string
        rows:
          type: array
          items:
            type: array
            items:
              description: A SQLite value (integer, number, string or null). BLOBs are returned as lowercase hex strings.
        changes:
          type: integer
          format: int64
          description: Rows inserted, updated or deleted by the statement.
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
          description: HTTP status code.
        message:
          type: string
          description: Human-readable error message.
        code:
          type: integer
          description: |
            SQLite extended result code, when the error comes from SQLite (e.g. 2067 for SQLITE_CONSTRAINT_UNIQUE, 787 for SQLITE_CONSTRAINT_FOREIGNKEY).